fix: graph css not distributed correctly
