+++ examples/react-app/src/configs.ts
@@ -35,8 +35,8 @@ export const entityQueries: TwinMakerQuery[] = [
   },
 ];
 export const dataBindingTemplate: IDataBindingTemplate = {
-  'sel_entity': (entityQueries[0] as TwinMakerEntityHistoryQuery).entityId
-}
+  sel_entity: (entityQueries[0] as TwinMakerEntityHistoryQuery).entityId,
+};
 
 // Video Player
 
@@ -44,24 +44,24 @@ export const dataBindingTemplate: IDataBindingTemplate = {
  * Simple Mode
  * Specify the video stream name to stream video directly from Kinesis Video Streams
  */
- export const kvsStreamName = '<your-kvs-stream-name>';
+export const kvsStreamName = '<your-kvs-stream-name>';
 
- /**
-  * AWS IoT TwinMaker Mode
-  * Specify videoEntityId and videoComponentName from AWS IoT TwinMaker workspace
-  */
- export const videoEntityId = undefined;
- export const videoComponentName = undefined;
+/**
+ * AWS IoT TwinMaker Mode
+ * Specify videoEntityId and videoComponentName from AWS IoT TwinMaker workspace
+ */
+export const videoEntityId = undefined;
+export const videoComponentName = undefined;
 
- /**
-  * Specify time range for the video player
-  */
- export const videoViewport: Viewport = {
-   // Live video playback - use duration
-   duration: '0',
-   // On-Demand video playback - specify start and end time for video
-   /*
+/**
+ * Specify time range for the video player
+ */
+export const videoViewport: Viewport = {
+  // Live video playback - use duration
+  duration: '0',
+  // On-Demand video playback - specify start and end time for video
+  /*
    start: new Date('<your-start-time>'),
    end: new Date('<your-end-time>'),
    */
- };
+};
