+++ examples/react-app/src/pages/Home/index.tsx
@@ -9,25 +9,28 @@ import ViewportControls from '../../components/ViewPort/Controls';
 import { Container, Header, SpaceBetween } from '@cloudscape-design/components';
 
 import { TimeSync } from '@iot-app-kit/react-components';
+/*
+import { KnowledgeGraph } from '@iot-app-kit/react-components/dist/es/components/knowledge-graph'; // TODO: Replace this with the public export once we launch it.
+*/
 
 const ScenePage = () => {
   return (
     <AppLayout>
       <TimeSync group={'main'}>
         <DashboardManager>
-            <SpaceBetween size={'s'}>
-              <Container>
-                <ViewportControls />
-              </Container>
-              <Container header={<Header>Scene</Header>}>
-                <div style={{ height: '864px', position: 'relative' }}>
-                  <SceneViewer />
-                </div>
-              </Container>
-              <Container header={<Header>Video Player</Header>}>
-                  <VideoPlayer />
-              </Container>
-            </SpaceBetween>
+          <SpaceBetween size={'s'}>
+            <Container>
+              <ViewportControls />
+            </Container>
+            <Container header={<Header>Scene</Header>}>
+              <div style={{ height: '864px', position: 'relative' }}>
+                <SceneViewer />
+              </div>
+            </Container>
+            <Container header={<Header>Video Player</Header>}>
+              <VideoPlayer />
+            </Container>
+          </SpaceBetween>
         </DashboardManager>
       </TimeSync>
     </AppLayout>
