+++ packages/react-components/.storybook/preview.js
@@ -1,9 +1,8 @@
-import '../styles.css';
-
 import { useDarkMode } from 'storybook-dark-mode';
 import { applyMode, Mode } from '@cloudscape-design/global-styles';
 import { useEffect } from 'react';
 
+import '@iot-app-kit/components/dist/iot-app-kit-components/iot-app-kit-components.css';
 import './global.scss';
 
 // https://storybook.js.org/docs/react/writing-stories/parameters#global-parameters
