+++ packages/react-components/package.json
@@ -16,12 +16,14 @@
   },
   "version": "5.8.1",
   "scripts": {
-    "start": "npx start-storybook -p 6007",
+    "start": "npm run build && npx start-storybook -p 6007",
     "dev": "npm-watch build",
     "clean": "rm -rf dist && rm -rf storybook-static && rm -rf test-results && rm -rf coverage",
-    "build": "npm run clean && npm run build:es && npm run build:cjs && npm run copy:styles",
-    "build:es": "tsc -p ./tsconfig.build.json --outDir ./dist/es",
-    "build:cjs": "tsc -p ./tsconfig.build.json --outDir ./dist/cjs --module CommonJS",
+    "build": "npm run clean && npm run build:es && npm run build:cjs",
+    "build:es": "tsc -p ./tsconfig.build.json --outDir ./dist/es && npm run copy-assets:es",
+    "build:cjs": "tsc -p ./tsconfig.build.json --outDir ./dist/cjs --module CommonJS && npm run copy-assets:cjs",
+    "copy-assets:es": "copyfiles -e \"**/*.tsx\" -e \"**/*.ts\" -e \"**/*.snap\" -e \"**/*.js\" -e \"**/*.jsx\" -e \"**/*.json\" \"src/**/*\" -u 1 dist/es/",
+    "copy-assets:cjs": "copyfiles -e \"**/*.tsx\" -e \"**/*.ts\" -e \"**/*.snap\" -e \"**/*.js\" -e \"**/*.jsx\" -e \"**/*.json\" \"src/**/*\" -u 1 dist/cjs/",
     "copy:styles": "cp ../components/styles.css styles.css",
     "test": "npm run test:jest && npm run test:types",
     "test:jest": "TZ=utc jest",
@@ -97,7 +99,6 @@
     "@iot-app-kit/core-util": "5.8.1",
     "@iot-app-kit/source-iottwinmaker": "5.8.1",
     "color": "^4.2.3",
-    "cytoscape": "3.25.0",
     "d3-array": "^3.2.3",
     "d3-shape": "^3.2.0",
     "dompurify": "3.0.3",
@@ -105,12 +106,13 @@
     "parse-duration": "^1.0.3",
     "react-cytoscapejs": "^2.0.0",
     "react-intl": "6.4.2",
+    "react-error-boundary": "^4.0.4",
     "styled-components": "^5.3.11",
     "uuid": "^9.0.0",
     "video.js": "8.3.0"
   },
   "peerDependencies": {
-    "cytoscape": "^3.23",
+    "cytoscape": "^3.25",
     "react": "^18",
     "react-dom": "^18"
   },
